rename dir to custom_...
